Refactor site dropdown: co-locate data with views
Move data hooks and handlers from the site-dropdown parent into MainView
and PublishPickerView so each view owns what it renders. The parent is
now a thin shell that holds only the main-vs-picker view state and the
menu trigger.

- MainView owns: start/stop, preview, push/pull, open folder/editor/terminal
- PublishPickerView owns: pickable-sites query, connect + push flow
- index.tsx owns: view state + Menu shell + DropdownTrigger

Uses `useIsMutating` with mutation keys in MainView so the Push button
reflects picker-initiated pushes too (the picker keeps its own
usePushSiteToLive instance for the post-connect push).

Also removes the one-line installed-apps.ts shared file — the
InstalledApps type is only consumed by apps/ui, so inlining it in the
data/core types file is simpler.
