--- tools/common/lib/user-settings/installed-apps.ts
@@ -1,4 +1,0 @@
-import type { SupportedEditor } from './editor';
-import type { SupportedTerminal } from './terminal';
-
-export type InstalledApps = Record< SupportedEditor | SupportedTerminal, boolean >;
